+++ src/DhtSensor.h
@@ -25,7 +25,7 @@ public:
 		this->newMeasurements = true;
 	}
 
-	void check(int currentTime = 0)
+	void check(unsigned long currentTime = 0)
 	{
 		// if more time than interval was spent
 		if (this->nextReadTime < currentTime)
@@ -56,7 +56,7 @@ public:
 private:
 	DHT* sensor;
 	bool newMeasurements = false;
-	int nextReadTime = 0;
+	unsigned long nextReadTime = 0;
 	int interval = 0;
 	float temperature = -999;
 	float humidty = -999;
